Is result a match

Decide if a results from a DOI query is a match or not.

is_result_match(
  query_title,
  result_title,
  query_orcid,
  result_orcid,
  query_authors,
  result_authors
)

Arguments

query_title
Title of the query reference

result_title
Title of the result reference

query_orcid
ORCiD of the query first author

result_orcid
ORCiD of the result first author

query_authors
Authors string for the query

result_authors
Authors string for the result

Value

Logical whether the result is a match or not

Details

Result is a match if the similarity between titles is sufficiently high or if the similarity is lower by the first author ORCiDs match or the first author names match.