Search DOI links

Search Crossref for potential linked references for a DOI

search_doi_links(
  doi,
  preprint = NULL,
  limit = 20,
  filter_matches = FALSE,
  verbose = TRUE
)

Arguments

doi
DOI to query

preprint
Logical. Whether doi is a preprint (TRUE) or publication (FALSE). If NULL (default) attempt to determine this from doi.

limit
Number of Crossref results to return. Default is 20, max is 1000.

filter_matches
Logical. If TRUE only return matches rather than all Crossref results.

verbose
Logical. Whether or not to print progress messages.

Value

tibble of potential links

Details

Based on the method described in "Day-to-day discovery of preprint–publication links" https://doi.org/10.1007/s11192-021-03900-7 and code from https://github.com/gcabanac/preprint-publication-linker. Query filters have been modified to allow reverse linking publications to preprints as well as preprints to publications.

Examples

search_doi_links("10.1101/133173")

#> Retrieving query data from Crossref...

#> Decided query is a preprint

#> Retrieving results data from Crossref...

#> Found 20 results

#> Checking results for matches...

#> Done!

#> # A tibble: 20 x 40
#>    alternative.id  container.title             created deposited published.print
#>    <chr>           <chr>                       <chr>   <chr>     <chr>          
#>  1 1305            Genome Biology              2017-0… 2020-10-… 2017-12        
#>  2 NA              PLOS Computational Biology  2018-0… 2019-10-… NA             
#>  3 2388            Genome Biology              2021-0… 2021-06-… 2021-12        
#>  4 NA              Development                 2019-0… 2021-04-… 2019-07-01     
#>  5 NA              Development                 2019-0… 2021-05-… NA             
#>  6 NA              Methods in Molecular Biolo… 2019-0… 2019-05-… 2019           
#>  7 NA              F1000Research               2020-0… 2020-06-… NA             
#>  8 615             Genome Medicine             2019-0… 2020-11-… 2019-12        
#>  9 1688            Genome Biology              2019-0… 2020-04-… 2019-12        
#> 10 10.1002/cpcb.86 Current Protocols in Cell … 2019-0… 2020-02-… 2019-06        
#> 11 NA              GigaScience                 2018-0… 2019-10-… 2018-05-01     
#> 12 NA              Bioinformatics              2021-0… 2021-06-… NA             
#> 13 7540            Nature Communications       2018-1… 2019-11-… 2018-12        
#> 14 2071            Genome Biology              2020-0… 2020-07-… 2020-12        
#> 15 1935            Genome Biology              2020-0… 2021-02-… 2020-12        
#> 16 10.7554/eLife.… eLife                       2019-0… 2020-05-… NA             
#> 17 NA              Development                 2019-0… 2021-04-… 2019-03-01     
#> 18 1303            Genome Biology              2017-0… 2017-08-… 2017-12        
#> 19 1284            Genome Biology              2017-0… 2020-10-… 2017-12        
#> 20 NA              GigaScience                 2018-0… 2019-10-… 2018-07-01     
#> # … with 35 more variables: published.online <chr>, doi <chr>, indexed <chr>,
#> #   issn <chr>, issue <chr>, issued <chr>, member <chr>, prefix <chr>,
#> #   publisher <chr>, score <chr>, source <chr>, reference.count <chr>,
#> #   references.count <chr>, is.referenced.by.count <chr>, title <chr>,
#> #   type <chr>, update.policy <chr>, url <chr>, volume <chr>, language <chr>,
#> #   short.container.title <chr>, author <list>, link <list>,
#> #   content_domain <list>, reference <list>, page <chr>, subject <chr>,
#> #   funder <list>, update_to <list>, license <list>, abstract <chr>,
#> #   assertion <list>, isbn <chr>, archive <chr>, match <lgl>